Skip index_bridging tests when there is no pageinspect
